Move set maximum number of threads test
