+++ core/src/main/resources/org/elasticsearch/bootstrap/test-framework.policy
@@ -56,7 +56,3 @@ grant codeBase "${codebase.junit-4.11.jar}" {
   // needed for TestClass creation
   permission java.lang.RuntimePermission "accessDeclaredMembers";
 };
-
-grant {
-  permission java.io.FilePermission "/proc/self/limits", "read";
-};
